pyproject.toml
[tool.poetry]
name = "qsgrc"
version = "0.1.0"
description = "QSG's 24hr of Lemons Race Computer Software for Car and Pit"
authors = ["Alex Hirschfeld <alex@d-qoi.com>"]
license = "GNU GPLv3"
readme = "README.md"
packages = [
    {include = "qsgrc/mono_car",from = "bases"},
    {include = "qsgrc/mono_pit",from = "bases"},
    {include = "qsgrc/RYLR896",from = "components"},
    {include = "qsgrc/alerts",from = "components"},
    {include = "qsgrc/event_server",from = "components"},
    {include = "qsgrc/log",from = "components"},
    {include = "qsgrc/messages",from = "components"},
    {include = "qsgrc/monitor",from = "components"},
    {include = "qsgrc/config",from = "components"},
]

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^14.2"
fastapi = {extras = ["standard"], version = "^0.115.8"}
pydantic-settings = "^2.7.1"
obd = "^0.7.2"
pyserial = "^3.5"
psutil = "^6.1.1"
pyserial-asyncio = "^0.6"
fastapi-sse = "^1.1.1"
nats-py = "^2.10.0"

[tool.poetry.requires-plugins]
poetry-multiproject-plugin = "^1.8.3"
poetry-polylith-plugin = "^1.36.2"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
black-macchiato = "^1.3.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
anyio = "^4.8.0"
pytest-timeout = "^2.3.1"
ipython = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
extraPaths = ["bases", "components"]

[tool.poetry.scripts]
build = "scripts.build_projects:build"
